+++ src/config/app.config.dev.ts
@@ -2,7 +2,6 @@ const devConfig = {
   baseUrl: 'https://api.exilence.de',
   production: false,
   sentryBrowserDsn: undefined,
-  trackingId: '',
   redirectUrl: 'https://api.exilence.de/api/authentication/redirect',
   oauthUrl: 'https://www.pathofexile.com',
   pathOfExileUrl: 'https://www.pathofexile.com',